requirements.txt
google-genai>=0.2.0
dashscope>=1.14.0
python-dateutil>=2.8.0
flask>=3.0.0
gunicorn>=21.0.0
